Don't use codespell's "rare" dictionary for CI spell check
By default, the "rare" dictionary is used in addition to the more conservative "clear" dictionary. The "rare" dictionary
contains common misspellings that are also valid spellings of less commonly used words. For example, the "rare"
dictionary contains "wan" as a misspelling of "want", but "wan" is a valid word (and, more relevant to the Arduino world,
also a valid acronym for "wide-area network").

The goal for the "Spell Check" workflow is to catch some typos without a significant number of false positives, even if
that means some misspellings slip through. I think the "clear" dictionary is more in line with that goal.